requirements.txt
# ─── Web Framework ────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ─── MongoDB ──────────────────────────────────────────────────
pymongo==4.6.1
motor==3.3.2                  # async MongoDB driver

# ─── AI / LLM ─────────────────────────────────────────────────
google-generativeai==0.8.4

# ─── Validation & Config ──────────────────────────────────────
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ─── Utilities ────────────────────────────────────────────────
python-dateutil==2.9.0
faker==23.2.1                 # seed fake data
rich==13.7.0                  # pretty CLI output
httpx==0.27.0                 # async HTTP client

# ─── Security ─────────────────────────────────────────────────
slowapi==0.1.9                # rate limiting for FastAPI
python-jose[cryptography]==3.3.0
